requirements.txt
uvicorn[standard]
fastapi
python-dotenv